Change identifier column to bigint.
The probe identifier is an 8-byte integer. See
https://github.com/freenet/fred-official/blob/build01444/src/freenet/node/probe/Probe.java#L303

+++ fnprobe/db.py
@@ -141,7 +141,7 @@ class Database:
                      time       TIMESTAMP WITH TIME ZONE,
                      duration   INTERVAL,
                      htl        INTEGER,
-                     identifier INTEGER,
+                     identifier BIGINT,
                      percent    INTEGER
                     )""")
         cur.execute("""
